Move shapes into their own directory
And very carefully update the various require statements.
