+++ shapes/line.js
@@ -1,6 +1,6 @@
 var Fabric = require('fabric').fabric;
 var isNode = typeof window == 'undefined';
-var mixin = require('./src/mixin');
+var mixin = require('../src/mixin');
 
 var Line = Fabric.util.createClass(Fabric.Line, {
    type: 'line',
@@ -17,11 +17,11 @@ var Line = Fabric.util.createClass(Fabric.Line, {
 });
 
 var proto = Line.prototype;
-mixin(proto, require('./mixins/clone.mixin'));
-mixin(proto, require('./mixins/highlighted_stroke.mixin'));
-mixin(proto, require('./mixins/limit_size'));
-mixin(proto, require('./mixins/nudge'));
-mixin(proto, require('./mixins/two_point_interactivity'));
+mixin(proto, require('../mixins/clone.mixin'));
+mixin(proto, require('../mixins/highlighted_stroke.mixin'));
+mixin(proto, require('../mixins/limit_size'));
+mixin(proto, require('../mixins/nudge'));
+mixin(proto, require('../mixins/two_point_interactivity'));
 
 // Overrides the method from the to_object.mixin beccause the
 // constructor of Line is differen than the rest of the shapes.
